+++ embassy-sync/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+- Add `len`, `is_empty` and `is_full` functions to `Channel`.
+
 ## 0.5.0 - 2023-12-04
 
 - Add a PriorityChannel.
@@ -34,7 +38,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - impl `Default` for `Signal`
 - Remove unnecessary uses of `atomic-polyfill`
 - Add `#[must_use]` to all futures.
-
 
 ## 0.1.0 - 2022-08-26